refactor: rename local mcp package
