--- mcp/__init__.py
@@ -1,1 +1,0 @@
-# mcp package for resources and tools
