+++ pokemon_mcp/__init__.py
@@ -1,0 +1,1 @@
+# Local MCP-like utilities for resources and tools
