+++ pokemon_mcp/tools.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from mcp.server import Tool
+from pokemon_mcp.server import Tool
 
 from core.repository import get_pokemon
 from core.turn_engine import simulate_battle
